Make character under cursor visible
Prior to this commit, it was possible (in some cases)
for the bg and fg colors at the cursor to be the same,
resulting in the cursor hiding the character underneath it.

(I personally only noticed this in gvim,
but thought it would be safer to explicitly set the color
in the terminal, as well.)

This commit fixes this issue by setting the fg/bg of the cursor
to the inverse of the fg/bg colors in normal mode (`hi Normal`).